Revert "feat(desktop,web): native-fidelity terminal wheel scrolling via custom xterm handler + kitty identity (#5639)"
This reverts commit bd717ad48c26542da97e584b0af9dd36c04c34b4.
