+++ packages/host-service/src/terminal/env.test.ts
@@ -419,8 +419,8 @@ describe("buildV2TerminalEnv", () => {
 		const env = buildV2TerminalEnv(baseParams);
 		expect(env).toMatchObject({
 			TERM: "xterm-256color",
-			TERM_PROGRAM: "kitty",
-			TERM_PROGRAM_VERSION: "0.42.0",
+			TERM_PROGRAM: "vscode",
+			TERM_PROGRAM_VERSION: "1.128.0",
 			COLORTERM: "truecolor",
 			PWD: "/tmp/workspace",
 			SUPERSET_TERMINAL_ID: "term-1",
@@ -431,7 +431,7 @@ describe("buildV2TerminalEnv", () => {
 			SUPERSET_AGENT_HOOK_PORT: "51741",
 			SUPERSET_AGENT_HOOK_VERSION: "2",
 		});
-		expect(env.TERM_PROGRAM).toBe("kitty");
+		expect(env.TERM_PROGRAM).toBe("vscode");
 		expect(env.SHELL).toBe("/bin/zsh");
 		expect(env.LANG).toContain("UTF-8");
 	});
