--- packages/shared/src/terminal-wheel-handler/index.ts
@@ -1,13 +1,0 @@
-export { type CellDimensions, getCellDimensions } from "./cell-dimensions";
-export {
-	buildArrowSequence,
-	buildSgrWheelReport,
-	consumeWheelLines,
-	createSgrMouseModeTracker,
-	createTerminalWheelEventHandler,
-	createWheelLineState,
-	installTerminalWheelEventHandler,
-	type WheelLineContext,
-	type WheelLineInput,
-	type WheelLineState,
-} from "./terminal-wheel-handler";
